requirements.txt
torch
fastapi
uvicorn
torchvision
pillow
numpy
python-multipart
metaflow
python-dotenv==1.0.1
split-folders
torch-summary
pandas
seaborn
opencv-python
tabulate # for pandas
